Added a note about the number of registered devices in "Disclaimer" section.
The section was reorganized in 3 notes.

The High Level Data Fusion component retrieves the number of registered wristbands from the GOST database. If you have a certain number of wristband registered, the system will attend message from the same number of devices (i.e.: if you register 1000 wristaband, you must execute a test with 1000 wristband. To reduce the number of devices it is mandatory to delete the GOST database and register again the wristbands).